Refactor AssetViewer
