+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/AssetViewerPreview.tsx
