+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/AssetViewerUpload.tsx
