+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/AssetViewerUtils.tsx
